+++ ReboundPatches.cs
@@ -17,11 +17,9 @@ namespace Rebound
     {
         [HarmonyPrefix]
         [HarmonyPatch(nameof(Player.HandleJump))]
-        public static bool HandleJumpPrefix_HandleRebound(Player __instance)
+        public static bool HandleJumpPrefix_TriggerRebound(Player __instance)
         {
             if (!__instance == ReboundPlugin.player) { return true; }
-
-            //RBSettings.SetSettingsInPlugin(); // allow live config update
             
             __instance.jumpedThisFrame = false;
             __instance.timeSinceJumpRequested += Reptile.Core.dt;
@@ -46,13 +44,16 @@ namespace Rebound
 
         [HarmonyPrefix]
         [HarmonyPatch(nameof(Player.OnLanded))]
-        public static bool OnLandedPrefix_ReboundVel(Player __instance) {
+        public static bool OnLandedPrefix_ReboundDistance(Player __instance) {
             if (!__instance == ReboundPlugin.player) { return true; }
+            
+            __instance.boostpackTrailDefaultTime = ReboundPlugin.originalTrailTime;
+            __instance.boostpackTrailDefaultWidth = ReboundPlugin.originalTrailWidth;
             
             ReboundPlugin.landTime = 0f;
             ReboundPlugin.distanceFallenFromPeakOfJump += __instance.reallyMovedVelocity.y; // adds "velocity" from frame before that probably isn't caught. make sure this works the way we want it to!
             //ReboundPlugin.groundedPositionY = __instance.tf.position.y;
-            if (__instance.comboTimeOutTimer <= 0) { ReboundPlugin.CancelRebound(); }
+            //if (__instance.comboTimeOutTimer <= 0) { ReboundPlugin.CancelRebound(); }
             return true;
         }
 
@@ -74,18 +75,24 @@ namespace Rebound
         [HarmonyPatch(nameof(Player.LandCombo))]
         public static bool LandComboPrefix_ExtendForRebound(Player __instance) {
             if (!__instance == ReboundPlugin.player) { return true; }
-            
-            return (!(ReboundPlugin.CanRebound() && !__instance.slideAbility.stopDecided));
+
+            // don't land combo if the player can still rebound
+            return !ReboundPlugin.CanRebound(); //(!(ReboundPlugin.CanRebound() && !__instance.slideAbility.stopDecided));
         }
 
         [HarmonyPrefix]
         [HarmonyPatch(nameof(Player.ActivateAbility))]
-        public static bool AbilityPrefix_CancelRBIfSlide(Ability a, Player __instance) {
+        public static bool AbilityPrefix_CancelRebound(Ability a, Player __instance) {
             if (!__instance == ReboundPlugin.player) { return true; }
 
-            if (__instance.ability == null && a == __instance.slideAbility && ReboundPlugin.CanRebound() && ReboundPlugin.landTime > Reptile.Core.dt*3f) {
+            __instance.boostpackTrailDefaultTime = ReboundPlugin.originalTrailTime;
+            __instance.boostpackTrailDefaultWidth = ReboundPlugin.originalTrailWidth;
+
+            if ((__instance.ability == null || __instance.ability == __instance.slideAbility) && a != __instance.boostAbility && 
+            ReboundPlugin.CanRebound() && ReboundPlugin.landTime > Reptile.Core.dt*3f) {
                 ReboundPlugin.CancelRebound();
             }
+            
             return true;
         }
     }