+++ ReboundPlugin.cs
@@ -23,7 +23,7 @@ namespace Rebound
         public static bool MPlusActive = false; 
 
         public static float landTime = float.PositiveInfinity;
-        public static Vector2 landingVelocity = new Vector2 (0, 0);
+        public static Vector2 landingVelocity = new Vector2(0, 0);
         public static float distanceFallenFromPeakOfJump = 0f;
         public static float landingComboMeter = 1f;
 
@@ -44,13 +44,18 @@ namespace Rebound
         public static bool tempDisableBoostAfterRebound = true;
 
         // non-configurable values, must be set here (New Years Update)
-        public const float comboCost = 0.25f; // negative = increase
+        public const float comboCost = 0.215f; // negative = increase
         public const float boostCost = 0f;
         public const bool allowBoostedRebounds = true;
-        //public static bool canCancelCombo = false; // buggy and unfinished. intended more for vanilla
+        public static bool canCancelCombo = true; // buggy and unfinished. intended more for vanilla
 
         public static List<string> cancelReboundActions = new List<string>();
         public static List<string> doReboundActions = new List<string>();
+        
+        public static float reboundTrailTime { get { return RBSettings.config_trailLength.Value + 0.000125f; } }
+        public static float reboundTrailWidth { get { return RBSettings.config_trailWidth.Value + 0.000125f; } }
+        public static float originalTrailTime; 
+        public static float originalTrailWidth;
 
         private void Awake()
         {
@@ -78,19 +83,26 @@ namespace Rebound
             if (landTime < maxLandingTimeToRebound) { landTime += Reptile.Core.dt; } 
             
             if (!player.IsGrounded()) {
-                landingVelocity = new Vector2 (player.GetForwardSpeed(), player.GetVelocity().y);
+                landingVelocity = new Vector2(player.GetForwardSpeed(), player.GetVelocity().y);
                 ReboundPlugin.landingComboMeter = player.comboTimeOutTimer;
 
                 if (player.GetVelocity().y < 0) {
                     distanceFallenFromPeakOfJump += Mathf.Abs(player.GetVelocity().y);
                     landTime = 0f;
-                } else {
-                    distanceFallenFromPeakOfJump = 0; // no longer the peak!
-                }
-
+                } else { distanceFallenFromPeakOfJump = 0; } // no longer the peak!
 
             } else if (player.ability == player.boostAbility && landTime >= maxLandingTimeToRebound && !MPlusActive) {
                 player.LandCombo();
+            }
+
+            if (player.ability == player.slideAbility) { landTime = float.PositiveInfinity; }
+
+            if (player.boostpackTrailDefaultTime == reboundTrailTime && (player.boostTrailTimer == 0f)) {
+                player.boostpackTrailDefaultTime = originalTrailTime;
+                player.boostpackTrailDefaultWidth = originalTrailWidth;
+            } else if (player.boostpackTrailDefaultWidth == reboundTrailWidth) {
+                player.boostTrailTimer += Mathf.Min(Core.dt, player.GetVelocity().y/500f)*6f; 
+                player.trailWidth += Mathf.Min(Core.dt, player.GetVelocity().y/500f); 
             }
         }
 
@@ -100,7 +112,7 @@ namespace Rebound
 
         public static void ReboundTrick() {
             bool isBoosting = allowBoostedRebounds ? player.boosting : false;
-            float floorAngle = (Vector3.Dot(Vector3.ProjectOnPlane(player.motor.groundNormalVisual, Vector3.up).normalized, player.dir));
+            float floorAngle = Vector3.Dot(Vector3.ProjectOnPlane(player.motor.groundNormalVisual, Vector3.up).normalized, player.dir);
             
             // Force rebound animations to properly transition to jump/fall animations
             player.animInfosSets[(int)player.moveStyle][Animator.StringToHash("jumpTrick1")].fadeTo[player.fallHash] = 1f;
@@ -125,61 +137,71 @@ namespace Rebound
                     player.PlayAnim(Animator.StringToHash("jumpTrick1"), true, true, -1f);
                 }
                 
-                string normalTrickName = player.moveStyle == MoveStyle.BMX ? "360 Backflip" : (player.moveStyle == MoveStyle.SKATEBOARD ? "McTwist" : "Corkscrew");
+                string normalTrickName = player.moveStyle == MoveStyle.BMX ? "360 Backflip" : 
+                    (player.moveStyle == MoveStyle.SKATEBOARD ? "McTwist" : "Corkscrew");
                 player.DoTrick(Player.TrickType.AIR, "Rebound " + normalTrickName, 0); 
             }
 
             // Counteract M+ fast fall exploitation - bounceCap should never be run into in vanilla
-            float targetHeight = (distanceFallenFromPeakOfJump*reboundVelocityMultiplier); // how high a rebound SHOULD be taking you
+            float targetHeight = distanceFallenFromPeakOfJump * reboundVelocityMultiplier; // how high a rebound SHOULD be taking you
             float reboundCalculatedByHeight = Mathf.Sqrt(2f * targetHeight * player.motor.gravity); // velocity to get to said height
-            float bounceCap = (reboundCalculatedByHeight/9f) + heightCapGenerosity; // why divide by nine? i dunno, but it works!
+            float bounceCap = (reboundCalculatedByHeight/9f) + heightCapGenerosity; // why divide by nine? i dunno, but it works well!
 
-            if (!alwaysCalculateBasedOnHeight && capBasedOnHeight) {
-                reboundVelocity.y = Mathf.Min(reboundVelocity.y, bounceCap);
-            } else if (alwaysCalculateBasedOnHeight) {
-                reboundVelocity.y = bounceCap - heightCapGenerosity;
-            }
+            if (alwaysCalculateBasedOnHeight) { reboundVelocity.y = bounceCap - heightCapGenerosity; }
+            else if (capBasedOnHeight) { reboundVelocity.y = Mathf.Min(reboundVelocity.y, bounceCap); }
             
             // Rebound off launcher bonus
             if (player.onLauncher) { 
                 if (!RBSettings.config_slopeOnLauncher.Value) { floorAngle = 0f; }
-                float launcherBonus = player.onLauncher.parent.gameObject.name.Contains("Super") ? player.jumpSpeedLauncher * 1.4f : player.jumpSpeedLauncher;
-                reboundVelocity.y = Mathf.Max(reboundVelocity.y + (launcherBonus*launcherBonusMultiplier), launcherBonus); //launcher bonus multiplier should be configurable
+                float launcherBonus = player.onLauncher.parent.gameObject.name.Contains("Super") ? 
+                    player.jumpSpeedLauncher * 1.4f : player.jumpSpeedLauncher;
+                reboundVelocity.y = Mathf.Max(reboundVelocity.y + (launcherBonus*launcherBonusMultiplier), launcherBonus);
             }
 
             // Constrain to min and max
             reboundVelocity.y = Mathf.Max(reboundVelocity.y, minReboundYVelocity);
-
             if (reboundVelocity.y > maxReboundYVelocity && maxReboundYVelocity != 0) {
                 reboundVelocity.x = maxReboundYVelocity * (reboundVelocity.x/reboundVelocity.y); // reduce speed to follow same arc
                 reboundVelocity.y = maxReboundYVelocity;
             }
 
             // Adjust velocity based on floor normal
-            Vector2 floorAngleXY = new Vector2 (Mathf.Sin(floorAngle)*(RBSettings.config_slopeMultiplierX.Value), Mathf.Cos(floorAngle)*(RBSettings.config_slopeMultiplierY.Value));
+            Vector2 floorAngleXY = new Vector2(
+                Mathf.Sin(floorAngle)*(RBSettings.config_slopeMultiplierX.Value), 
+                Mathf.Cos(floorAngle)*(RBSettings.config_slopeMultiplierY.Value));
             if (floorAngleXY.x < 0f) { floorAngleXY.y += 1f; } // higher instead of lower
             //floorAngleXY.y = Mathf.Lerp(floorAngleXY.y, 1f, 0.5f);
 
             player.motor.SetVelocityYOneTime(reboundVelocity.y * floorAngleXY.y);
             float newForwardSpeed = reboundVelocity.x + (reboundVelocity.y * floorAngleXY.x);
             player.SetForwardSpeed(newForwardSpeed);
-            if (newForwardSpeed < 0f) {
-                player.SetRotation(-player.dir);
-            }
+            if (newForwardSpeed < 0f) { player.SetRotation(-player.dir); } // turn player around if going backwards
             
             // Handle combo meter
-            //float oldComboTimer = player.comboTimeOutTimer;
             if (comboCost != 0f) { 
-                if ((player.comboTimeOutTimer - comboCost) > 1f) {
-                    player.ResetComboTimeOut();
-                } else {
+                if ((player.comboTimeOutTimer - comboCost) > 1f) { player.ResetComboTimeOut(); } 
+                else {
+                    //float oldComboTime = player.comboTimeOutTimer;
                     player.DoComboTimeOut(comboCost);
+                    //bool withinBuffer = (oldComboTime > (comboCost*0.6f) && player.comboTimeOutTimer < 0f) || player.comboTimeOutTimer == 0f;
                     player.comboTimeOutTimer = Mathf.Clamp(player.comboTimeOutTimer, 0f, 1f); 
+                        //Mathf.Clamp(player.comboTimeOutTimer, (withinBuffer ? 0.000001f : 0f), 1f); 
                 }
             }
-
+            
             if (player.comboTimeOutTimer == 0f) { player.LandCombo(); }
             //if (boostCost != 0f) { player.AddBoostCharge(-boostCost); }
+
+            // Extend trail effect
+            if (player.boostpackTrailDefaultTime < reboundTrailTime) {
+                originalTrailTime = player.boostpackTrailDefaultTime; 
+                originalTrailWidth = player.boostpackTrailDefaultWidth;
+                player.boostpackTrailDefaultTime = reboundTrailTime;
+                player.boostpackTrailDefaultWidth = reboundTrailWidth;
+            } 
+            
+            player.boostTrailTimer = player.boostpackTrailDefaultTime;
+            player.trailWidth = player.boostpackTrailDefaultWidth;  
         }
 
         public static void PrepForRebound() {
@@ -197,7 +219,8 @@ namespace Rebound
             player.ForceUnground(true);
             player.radialHitbox.SetActive(true);
             
-            player.AudioManager.PlaySfxGameplay(SfxCollectionID.GenericMovementSfx, AudioClipID.jump_special, player.playerOneShotAudioSource, 0f);
+            player.AudioManager.PlaySfxGameplay(SfxCollectionID.GenericMovementSfx, AudioClipID.jump_special, 
+                player.playerOneShotAudioSource, 0f);
             player.PlayVoice(AudioClipID.VoiceJump, VoicePriority.MOVEMENT, true);
             if (!isBoosting) { player.PlayAnim(Animator.StringToHash("jumpTrick1"), true, false, -1f); }
             else { player.PlayAnim(Animator.StringToHash("jump"), true, false, -1f); }
@@ -207,27 +230,27 @@ namespace Rebound
         }
 
         public static void CancelRebound() {
-            if (!ComboCanBeCancelled()) {return;}
+            if (!ComboCanBeCancelled() || !RBSettings.config_preventComboExtend.Value) { return; }
             ReboundPlugin.landTime = ReboundPlugin.maxLandingTimeToRebound + 1f;
             player.LandCombo();
         }
 
-        public static bool ComboCanBeCancelled() {
+        public static bool ComboCanBeCancelled() { 
+            if (!canCancelCombo) { return false; }
             bool mpluscombodone = MPlusActive ? player.comboTimeOutTimer <= 0 : true;
-            return mpluscombodone && player.ability != player.slideAbility && !player.IsPerformingManualTrick && !player.slideAbility.locked
-            && player.ability != player.ledgeClimbAbility && player.switchMoveStyleAbility != player.ability;
+            return mpluscombodone && player.ability != player.slideAbility && //&& !player.IsPerformingManualTrick
+            !player.slideAbility.locked && player.ability != player.ledgeClimbAbility && player.switchMoveStyleAbility != player.ability;
         }
 
         private void OnDestroy() { Harmony.UnpatchSelf(); }
 
         public static bool ActionsPressed(List<string> actions, bool checkIfAllPressed) {
-            if (actions?.Any() != true) {return false;}
+            if (actions?.Any() != true) { return false; }
             bool returnValue = checkIfAllPressed;
             foreach (string actionPropertyOrMethod in actions) {
                 if (TranslateActionString(actionPropertyOrMethod) == !checkIfAllPressed) 
                 { returnValue = !checkIfAllPressed; }
-            }
-            return returnValue;      
+            } return returnValue;      
         }
 
         public static bool TranslateActionString(string action) {